+++ vnext/Microsoft.ReactNative/Fabric/Composition/CompositionEventHandler.cpp
@@ -117,7 +117,6 @@ struct CompositionKeyboardSource
   CompositionEventHandler *m_outer{nullptr};
 };
 
-#ifdef USE_WINUI3
 struct CompositionInputKeyboardSource : winrt::implements<
                                             CompositionInputKeyboardSource,
                                             winrt::Microsoft::ReactNative::Composition::Input::KeyboardSource> {
@@ -137,7 +136,6 @@ struct CompositionInputKeyboardSource : winrt::implements<
  private:
   winrt::Microsoft::UI::Input::InputKeyboardSource m_source{nullptr};
 };
-#endif
 
 CompositionEventHandler::CompositionEventHandler(
     const winrt::Microsoft::ReactNative::ReactContext &context,
@@ -145,7 +143,6 @@ CompositionEventHandler::CompositionEventHandler(
     : m_context(context), m_wkRootView(reactNativeIsland) {}
 
 void CompositionEventHandler::Initialize() noexcept {
-#ifdef USE_WINUI3
   if (auto island = m_wkRootView.get().Island()) {
     auto pointerSource = winrt::Microsoft::UI::Input::InputPointerSource::GetForIsland(island);
 
@@ -323,11 +320,9 @@ void CompositionEventHandler::Initialize() noexcept {
           }
         });
   }
-#endif
 }
 
 CompositionEventHandler::~CompositionEventHandler() {
-#ifdef USE_WINUI3
   if (auto strongRootView = m_wkRootView.get()) {
     if (auto island = strongRootView.Island()) {
       auto pointerSource = winrt::Microsoft::UI::Input::InputPointerSource::GetForIsland(island);
@@ -342,7 +337,6 @@ CompositionEventHandler::~CompositionEventHandler() {
       keyboardSource.CharacterReceived(m_characterReceivedToken);
     }
   }
-#endif
 
   if (m_hcursorOwned) {
     ::DestroyCursor(m_hcursor);