+++ vnext/Microsoft.ReactNative/Utils/KeyboardUtils.cpp
@@ -5,9 +5,7 @@
 
 #include <winrt/Windows.UI.Core.h>
 
-#ifdef USE_WINUI3
 #include <winrt/Microsoft.UI.Input.h>
-#endif
 
 namespace Microsoft::ReactNative {
 
@@ -374,14 +372,7 @@ std::string FromVirtualKey(winrt::Windows::System::VirtualKey virtualKey, bool f
 }
 
 bool IsModifiedKeyPressed(winrt::CoreWindow const &coreWindow, winrt::Windows::System::VirtualKey virtualKey) {
-#ifndef USE_WINUI3
-  if (!coreWindow) {
-    return GetKeyState(static_cast<int>(virtualKey)) < 0;
-  }
-  auto const &keyState = coreWindow.GetKeyState(virtualKey);
-#else
   auto const &keyState = winrt::Microsoft::UI::Input::InputKeyboardSource::GetKeyStateForCurrentThread(virtualKey);
-#endif // USE_WINUI3
   return (keyState & winrt::CoreVirtualKeyStates::Down) == winrt::CoreVirtualKeyStates::Down;
 }