+++ vnext/Shared/Hermes/HermesRuntimeTargetDelegate.cpp
@@ -15,6 +15,7 @@
 
 #include "HermesRuntimeTargetDelegate.h"
 #include <jsinspector-modern/RuntimeTarget.h>
+#include <windows.h>
 #include <utility>
 #include "HermesRuntimeAgentDelegate.h"
 
@@ -268,7 +269,10 @@ tracing::RuntimeSamplingProfile HermesRuntimeTargetDelegate::collectSamplingProf
 
   // Return the complete profile with samples. Wrap the raw profile since it owns the strings.
   return tracing::RuntimeSamplingProfile(
-      "Hermes", 1234, std::move(readerState.samples), std::make_unique<HermesRawRuntimeProfile>(std::move(profile)));
+      "Hermes",
+      GetCurrentProcessId(),
+      std::move(readerState.samples),
+      std::make_unique<HermesRawRuntimeProfile>(std::move(profile)));
 }
 
 } // namespace Microsoft::ReactNative