+++ packages/@react-native-windows/automation/CHANGELOG.md
@@ -1,8 +1,105 @@
 # Change Log - @react-native-windows/automation
 
-<!-- This log was last generated on Tue, 18 Nov 2025 06:41:41 GMT and should not be manually modified. -->
+<!-- This log was last generated on Sat, 13 Dec 2025 06:33:40 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.0.0-canary.1020
+
+Sat, 13 Dec 2025 06:33:40 GMT
+
+### Changes
+
+- Bump @react-native-windows/automation-channel to v0.0.0-canary.1020
+
+## 0.0.0-canary.1019
+
+Sun, 07 Dec 2025 06:34:40 GMT
+
+### Changes
+
+- Bump @react-native-windows/automation-channel to v0.0.0-canary.1019
+
+## 0.0.0-canary.1018
+
+Sat, 06 Dec 2025 06:38:46 GMT
+
+### Changes
+
+- Bump @react-native-windows/automation-channel to v0.0.0-canary.1018
+
+## 0.0.0-canary.1017
+
+Fri, 05 Dec 2025 06:34:57 GMT
+
+### Changes
+
+- Bump @react-native-windows/automation-channel to v0.0.0-canary.1017
+
+## 0.0.0-canary.1016
+
+Wed, 03 Dec 2025 06:40:45 GMT
+
+### Changes
+
+- Use 0.0.0-canary.1015 version for public NPM packages (vmorozov@microsoft.com)
+- Bump @react-native-windows/automation-channel to v0.0.0-canary.1016
+
+## 0.3.437-0
+
+Tue, 02 Dec 2025 06:36:52 GMT
+
+### Changes
+
+- Bump @react-native-windows/automation-channel to v0.12.355-0
+
+## 0.3.436
+
+Mon, 01 Dec 2025 06:36:40 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.354
+
+## 0.3.435
+
+Thu, 27 Nov 2025 06:35:01 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.353
+
+## 0.3.434
+
+Wed, 26 Nov 2025 06:36:13 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.352
+
+## 0.3.433
+
+Fri, 21 Nov 2025 06:43:27 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.351
+
+## 0.3.432
+
+Thu, 20 Nov 2025 06:39:28 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.350
+
+## 0.3.431
+
+Wed, 19 Nov 2025 06:41:09 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.349
 
 ## 0.3.430